Partially complete score system. Added event for earthquake. Events are now synchornized on all clients. Misc bug fixes.

+++ Lords-of-Distortion/Assets/Scripts/PlayerStats.cs
@@ -4,7 +4,7 @@
 public class PlayerStats {
 	public int deaths = 0;
 	public int kills = 0;
-	public int score = 0;
+	public float score = 0f;
 
 	public float timeOfDeath = float.PositiveInfinity;
